refactor(node): isolate index worker process helpers
Extract process namespace ownership, slow-open heartbeat, RPC query trait
adapters, and coalesced wake/deadline scheduling into private worker modules.
Move five existing namespace/heartbeat tests alongside their owners and add
five scheduling regressions. Preserve the existing worker algorithms.

Related to #742 and #645. Rust formatting, compilation, Clippy and tests
remain unverified; this is a source-only draft review commit.
