+++ crates/node/src/txindex_worker/scheduling.rs
@@ -1,0 +1,69 @@
+//! Coalesced wake handling and bounded waits for the node-owned worker.
+//!
+//! Wake messages are hints; the revision determines whether the chain moved.
+//! A buffered batch keeps its original deadline across subsequent wakes.
+
+use std::time::{Duration, Instant};
+
+use crossbeam_channel::Receiver;
+
+use super::TxIndexRuntime;
+
+#[derive(Clone, Copy, Debug, Eq, PartialEq)]
+pub(super) enum BatchWait {
+    Woken,
+    Deadline,
+    Stopped,
+}
+
+pub(super) fn wait_for_revision_quiet(
+    runtime: &TxIndexRuntime,
+    wake_rx: &Receiver<()>,
+    quiet_period: Duration,
+    mut seen_revision: u64,
+) -> Option<u64> {
+    loop {
+        if runtime.should_stop() {
+            return None;
+        }
+        match wake_rx.recv_timeout(quiet_period) {
+            Ok(()) => seen_revision = runtime.revision(),
+            Err(crossbeam_channel::RecvTimeoutError::Timeout) => {
+                let current = runtime.revision();
+                if current == seen_revision {
+                    return Some(current);
+                }
+                seen_revision = current;
+            }
+            Err(crossbeam_channel::RecvTimeoutError::Disconnected) => return None,
+        }
+    }
+}
+/// Waits for a wake hint or the pending batch's original deadline.
+pub(super) fn wait_for_batch_deadline(
+    runtime: &TxIndexRuntime,
+    wake_rx: &Receiver<()>,
+    deadline: Instant,
+) -> BatchWait {
+    if runtime.should_stop() {
+        return BatchWait::Stopped;
+    }
+    let Some(remaining) = deadline.checked_duration_since(Instant::now()) else {
+        return BatchWait::Deadline;
+    };
+    if remaining.is_zero() {
+        return BatchWait::Deadline;
+    }
+    match wake_rx.recv_timeout(remaining) {
+        Ok(()) if runtime.should_stop() => BatchWait::Stopped,
+        Ok(()) => BatchWait::Woken,
+        Err(crossbeam_channel::RecvTimeoutError::Timeout) if runtime.should_stop() => {
+            BatchWait::Stopped
+        }
+        Err(crossbeam_channel::RecvTimeoutError::Timeout) => BatchWait::Deadline,
+        Err(crossbeam_channel::RecvTimeoutError::Disconnected) => BatchWait::Stopped,
+    }
+}
+
+#[cfg(test)]
+mod tests;
