+++ crates/node/src/txindex_worker_lifecycle_tests.rs
@@ -100,67 +100,6 @@ fn query_adapter_returns_unavailable_for_shutdown_abandoned() {
 }
 
 #[test]
-fn namespace_registry_claims_and_releases() {
-    let registry = NamespaceRegistry::new();
-    let key = PathBuf::from("/tmp/test-namespace-claim");
-
-    // First claim succeeds.
-    assert!(registry.claim(key.clone(), 1));
-    // Second claim by different owner fails.
-    assert!(!registry.claim(key.clone(), 2));
-    // Release by wrong owner does nothing.
-    registry.release(&key, 2);
-    assert!(matches!(
-        registry.entries.lock().get(&key),
-        Some(NamespaceEntry::Active(1))
-    ));
-    // Release by correct owner removes the entry.
-    registry.release(&key, 1);
-    assert!(registry.entries.lock().get(&key).is_none());
-}
-
-#[test]
-fn namespace_registry_poisons_on_abandon() {
-    let registry = NamespaceRegistry::new();
-    let key = PathBuf::from("/tmp/test-namespace-poison");
-
-    assert!(registry.claim(key.clone(), 1));
-    registry.poison(&key, 1);
-    assert!(registry.is_poisoned(&key));
-    // Poisoned namespace cannot be claimed.
-    assert!(!registry.claim(key, 2));
-}
-
-#[test]
-fn namespace_registry_poison_only_for_matching_owner() {
-    let registry = NamespaceRegistry::new();
-    let key = PathBuf::from("/tmp/test-namespace-poison-owner");
-
-    assert!(registry.claim(key.clone(), 1));
-    // Poison by wrong owner does nothing.
-    registry.poison(&key, 2);
-    assert!(!registry.is_poisoned(&key));
-    assert!(matches!(
-        registry.entries.lock().get(&key),
-        Some(NamespaceEntry::Active(1))
-    ));
-}
-
-#[test]
-fn namespace_registry_validates_child() {
-    let root = Path::new("/tmp");
-    assert!(NamespaceRegistry::validate_child(root, "txindex").is_ok());
-    assert!(NamespaceRegistry::validate_child(root, "").is_err());
-    assert!(NamespaceRegistry::validate_child(root, ".").is_err());
-    assert!(NamespaceRegistry::validate_child(root, "..").is_err());
-    #[cfg(unix)]
-    {
-        assert!(NamespaceRegistry::validate_child(root, "foo/bar").is_err());
-        assert!(NamespaceRegistry::validate_child(root, "/absolute").is_err());
-    }
-}
-
-#[test]
 fn generation_revoke_makes_publication_noop() {
     let generation_tok = Generation::new(42);
     assert_eq!(generation_tok.id(), 42);
@@ -175,13 +114,4 @@ fn generation_clone_shares_revocation() {
     let gen_clone = generation_tok.clone();
     generation_tok.revoke();
     assert!(gen_clone.is_revoked(), "clone must see revocation");
-}
-
-#[test]
-fn heartbeat_starts_and_stops() {
-    let heartbeat = Heartbeat::start("test", "test-ns".to_owned(), "test-backend".to_owned());
-    // Give it a moment to potentially emit.
-    std::thread::sleep(std::time::Duration::from_millis(50));
-    heartbeat.stop_and_join();
-    // If we get here without hanging, the heartbeat stopped and joined.
 }
